+++ src/lib.rs
@@ -16,6 +16,9 @@
 //! ```
 //!
 
+#[macro_use] extern crate lazy_static;
+extern crate regex;
+
 mod ftp;
 pub mod status;